// frontend/src/app/core/generated/circabc/model/user.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * a CIRCABC user
 */
export interface User { 
    userId?: string;
    firstname?: string;
    lastname?: string;
    email?: string;
    phone?: string;
    uiLang?: string;
    contentFilterLang?: string;
    visibility?: boolean;
    /**
     * properties field will receive all the configuration parameters relative to the user accound. For example notification / prefered language / etc. 
     */
    properties?: { [key: string]: string; };
    /**
     * Used to get the URL of the user picture that will be shown everywhere in the app 
     */
    avatar?: string;
    /**
     * Is this avatar the default avatar? 
     */
    defaultAvatar?: boolean;
}